+++ src/test/suite/modules/utils.test.ts
@@ -6,16 +6,16 @@ suite('utils.ts tests', () => {
 
   test('uniqBy test', () => {
     const list: any[] = [
-      { PackageName: "nugetpackagemanagergui", Versions: ["1", "2"] },
-      { PackageName: "SixLaborsCaptcha.Core", Versions: ["1", "2"] },
+      { packageName: "nugetpackagemanagergui", versions: ["1", "2"] },
+      { packageName: "SixLaborsCaptcha.Core", versions: ["1", "2"] },
     ];
-    const uniqList = uniqBy([...list, { PackageName: "nugetpackagemanagergui", Versions: ["1", "2"] }], "PackageName");
+    const uniqList = uniqBy([...list, { packageName: "nugetpackagemanagergui", versions: ["1", "2"] }], "packageName");
     assert.deepStrictEqual(list, uniqList);
   });
 
   test('jsonToQueryString test', () => {
-    const obj: any = { PackageName: "nugetpackagemanagergui", Versions: "12.22" };
+    const obj: any = { packageName: "nugetpackagemanagergui", versions: "12.22" };
     const query = jsonToQueryString(obj);
-    assert.deepStrictEqual(query, `?PackageName=nugetpackagemanagergui&Versions=12.22`);
+    assert.deepStrictEqual(query, `?packageName=nugetpackagemanagergui&versions=12.22`);
   });
 });